order csr

+++ src/main/java/com/training360/cafebabeswebshop/order/OrderService.java
@@ -22,10 +22,14 @@ public class OrderService {
     }
 
     public long saveOrderAndGetId(Authentication authentication, Order order){
-        long id = orderDao.saveOrderAndGetId(authentication.getName(), order);
-        addOrderedProducts(authentication, order);
-        basketDao.deleteBasket(authentication.getName());
-        return id;
+        if (basketDao.getBasketItems(authentication.getName()).size() > 0) {
+            long id = orderDao.saveOrderAndGetId(authentication.getName(), order);
+            addOrderedProducts(authentication, order);
+            basketDao.deleteBasket(authentication.getName());
+            return id;
+        } else {
+            throw new IllegalStateException("The basket is empty");
+        }
     }
 
     private void addOrderedProducts(Authentication authentication,Order order){
